[Native-Metadata] Rust implementation of DAO layer (#294)
* POC of native-metadata-jni


* Native DAO


* fix git action for native-metadata


* cleanup code


* fix after rebase main


---------

+++ lakesoul-common/src/main/java/com/dmetasoul/lakesoul/meta/dao/NamespaceDao.java
@@ -6,8 +6,10 @@ package com.dmetasoul.lakesoul.meta.dao;
 
 import com.dmetasoul.lakesoul.meta.DBConfig;
 import com.dmetasoul.lakesoul.meta.DBConnector;
-import com.dmetasoul.lakesoul.meta.DBUtil;
+import com.dmetasoul.lakesoul.meta.entity.JniWrapper;
 import com.dmetasoul.lakesoul.meta.entity.Namespace;
+import com.dmetasoul.lakesoul.meta.jnr.NativeMetadataJavaClient;
+import com.dmetasoul.lakesoul.meta.jnr.NativeUtils;
 import com.dmetasoul.lakesoul.meta.rbac.AuthZContext;
 import com.dmetasoul.lakesoul.meta.rbac.AuthZEnforcer;
 import dev.failsafe.internal.util.Lists;
@@ -17,10 +19,19 @@ import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Collections;
 import java.util.List;
+import java.util.stream.Collectors;
 
 public class NamespaceDao {
     public void insert(Namespace namespace) {
+        if (NativeUtils.NATIVE_METADATA_UPDATE_ENABLED) {
+            Integer count = NativeMetadataJavaClient.insert(
+                    NativeUtils.CodedDaoType.InsertNamespace,
+                    JniWrapper.newBuilder().addNamespace(namespace).build());
+            return;
+        }
         Connection conn = null;
         PreparedStatement pstmt = null;
         try {
@@ -40,6 +51,14 @@ public class NamespaceDao {
     }
 
     public Namespace findByNamespace(String name) {
+        if (NativeUtils.NATIVE_METADATA_QUERY_ENABLED) {
+            JniWrapper jniWrapper = NativeMetadataJavaClient.query(
+                    NativeUtils.CodedDaoType.SelectNamespaceByNamespace,
+                    Collections.singletonList(name));
+            if (jniWrapper == null) return null;
+            List<Namespace> namespaceList = jniWrapper.getNamespaceList();
+            return namespaceList.isEmpty() ? null : namespaceList.get(0);
+        }
         Connection conn = null;
         PreparedStatement pstmt = null;
         ResultSet rs = null;
@@ -61,6 +80,12 @@ public class NamespaceDao {
     }
 
     public void deleteByNamespace(String namespace) {
+        if (NativeUtils.NATIVE_METADATA_UPDATE_ENABLED) {
+            Integer count = NativeMetadataJavaClient.update(
+                    NativeUtils.CodedDaoType.DeleteNamespaceByNamespace,
+                    Collections.singletonList(namespace));
+            return;
+        }
         Connection conn = null;
         PreparedStatement pstmt = null;
         String sql = String.format("delete from namespace where namespace = '%s' ", namespace);
@@ -76,6 +101,14 @@ public class NamespaceDao {
     }
 
     public List<String> listNamespaces() {
+        if (NativeUtils.NATIVE_METADATA_QUERY_ENABLED) {
+            JniWrapper jniWrapper = NativeMetadataJavaClient.query(
+                    NativeUtils.CodedDaoType.ListNamespaces,
+                    Collections.emptyList());
+            if (jniWrapper == null) return null;
+            List<Namespace> namespaceList = jniWrapper.getNamespaceList();
+            return namespaceList.stream().map(Namespace::getNamespace).collect(Collectors.toList());
+        }
         Connection conn = null;
         PreparedStatement pstmt = null;
         ResultSet rs = null;
@@ -98,6 +131,11 @@ public class NamespaceDao {
     }
 
     public int updatePropertiesByNamespace(String namespace, String properties) {
+        if (NativeUtils.NATIVE_METADATA_UPDATE_ENABLED) {
+            return NativeMetadataJavaClient.update(
+                    NativeUtils.CodedDaoType.UpdateNamespacePropertiesByNamespace,
+                    Arrays.asList(namespace, properties));
+        }
         int result = 0;
         Connection conn = null;
         PreparedStatement pstmt = null;
@@ -136,7 +174,7 @@ public class NamespaceDao {
         return Namespace.newBuilder()
                 .setNamespace(rs.getString("namespace"))
                 .setProperties(rs.getString("properties"))
-                .setComment(comment == null ? "" : comment )
+                .setComment(comment == null ? "" : comment)
                 .setDomain(rs.getString("domain"))
                 .build();
     }
@@ -150,4 +188,3 @@ public class NamespaceDao {
                     .build();
 
 }
-
