+++ lakesoul-common/src/main/java/com/dmetasoul/lakesoul/meta/dao/TableNameIdDao.java
@@ -5,7 +5,10 @@
 package com.dmetasoul.lakesoul.meta.dao;
 
 import com.dmetasoul.lakesoul.meta.DBConnector;
+import com.dmetasoul.lakesoul.meta.entity.JniWrapper;
 import com.dmetasoul.lakesoul.meta.entity.TableNameId;
+import com.dmetasoul.lakesoul.meta.jnr.NativeMetadataJavaClient;
+import com.dmetasoul.lakesoul.meta.jnr.NativeUtils;
 import org.apache.commons.lang3.StringUtils;
 
 import java.sql.Connection;
@@ -13,11 +16,22 @@ import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Collections;
 import java.util.List;
+import java.util.stream.Collectors;
 
 public class TableNameIdDao {
 
     public TableNameId findByTableName(String tableName, String tableNamespace) {
+        if (NativeUtils.NATIVE_METADATA_QUERY_ENABLED) {
+            JniWrapper jniWrapper = NativeMetadataJavaClient.query(
+                    NativeUtils.CodedDaoType.SelectTableNameIdByTableName,
+                    Arrays.asList(tableName, tableNamespace));
+            if (jniWrapper == null) return null;
+            List<TableNameId> tableNameIdList = jniWrapper.getTableNameIdList();
+            return tableNameIdList.isEmpty() ? null : tableNameIdList.get(0);
+        }
         Connection conn = null;
         PreparedStatement pstmt = null;
         ResultSet rs = null;
@@ -39,12 +53,20 @@ public class TableNameIdDao {
         return tableNameId;
     }
 
-    public List<String> listAllNameByNamespace(String table_namespace) {
+    public List<String> listAllNameByNamespace(String tableNamespace) {
+        if (NativeUtils.NATIVE_METADATA_QUERY_ENABLED) {
+            JniWrapper jniWrapper = NativeMetadataJavaClient.query(
+                    NativeUtils.CodedDaoType.ListTableNameByNamespace,
+                    Collections.singletonList(tableNamespace));
+            if (jniWrapper == null) return null;
+            List<TableNameId> tableNameIdList = jniWrapper.getTableNameIdList();
+            return tableNameIdList.stream().map(TableNameId::getTableName).collect(Collectors.toList());
+        }
         Connection conn = null;
         PreparedStatement pstmt = null;
         ResultSet rs = null;
         String sql =
-                String.format("select table_name from table_name_id where table_namespace = '%s'", table_namespace);
+                String.format("select table_name from table_name_id where table_namespace = '%s'", tableNamespace);
         List<String> list = new ArrayList<>();
         try {
             conn = DBConnector.getConn();
@@ -63,6 +85,12 @@ public class TableNameIdDao {
     }
 
     public void insert(TableNameId tableNameId) {
+        if (NativeUtils.NATIVE_METADATA_UPDATE_ENABLED) {
+            Integer count = NativeMetadataJavaClient.insert(
+                    NativeUtils.CodedDaoType.InsertTableNameId,
+                    JniWrapper.newBuilder().addTableNameId(tableNameId).build());
+            return;
+        }
         Connection conn = null;
         PreparedStatement pstmt = null;
         try {
@@ -82,6 +110,10 @@ public class TableNameIdDao {
     }
 
     public void delete(String tableName, String tableNamespace) {
+        if (NativeUtils.NATIVE_METADATA_UPDATE_ENABLED) {
+            Integer count = NativeMetadataJavaClient.update(NativeUtils.CodedDaoType.DeleteTableNameIdByTableNameAndNamespace, Arrays.asList(tableName, tableNamespace));
+            return;
+        }
         Connection conn = null;
         PreparedStatement pstmt = null;
         String sql =
@@ -99,6 +131,12 @@ public class TableNameIdDao {
     }
 
     public void deleteByTableId(String tableId) {
+        if (NativeUtils.NATIVE_METADATA_UPDATE_ENABLED) {
+            Integer count = NativeMetadataJavaClient.update(
+                    NativeUtils.CodedDaoType.DeleteTableNameIdByTableId,
+                    Collections.singletonList(tableId));
+            return;
+        }
         Connection conn = null;
         PreparedStatement pstmt = null;
         String sql = String.format("delete from table_name_id where table_id = '%s' ", tableId);
